requirements.txt
tqdm==4.66.5
pydantic==2.8.2
langchain==0.2.10
langchain-community==0.2.9
langchain-core==0.2.22
langchain-openai==0.1.11
langchain-text-splitters==0.2.2
langchainhub==0.1.20
openai==1.40.2
six==1.16.0
oss2==2.18.4
playwright==1.45.1
beautifulsoup4==4.12.3
zhipuai==2.1.4.20230809.1
python-dotenv
fastapi==0.112.0
uvicorn==0.30.5
duckduckgo_search==6.2.6
pyOpenSSL
pyjwt